processor/requirements.txt
python-dotenv
pymongo
boto3
